Security Hub: Move some functions around.
