+++ internal/service/securityhub/find.go
@@ -94,80 +94,3 @@ func FindStandardsControlByStandardsSubscriptionARNAndStandardsControlARN(ctx co
 
 	return output, nil
 }
-
-func FindStandardsSubscription(ctx context.Context, conn *securityhub.SecurityHub, input *securityhub.GetEnabledStandardsInput) (*securityhub.StandardsSubscription, error) {
-	output, err := FindStandardsSubscriptions(ctx, conn, input)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if len(output) == 0 || output[0] == nil {
-		return nil, tfresource.NewEmptyResultError(input)
-	}
-
-	if count := len(output); count > 1 {
-		return nil, tfresource.NewTooManyResultsError(count, input)
-	}
-
-	return output[0], nil
-}
-
-func FindStandardsSubscriptions(ctx context.Context, conn *securityhub.SecurityHub, input *securityhub.GetEnabledStandardsInput) ([]*securityhub.StandardsSubscription, error) {
-	var output []*securityhub.StandardsSubscription
-
-	err := conn.GetEnabledStandardsPagesWithContext(ctx, input, func(page *securityhub.GetEnabledStandardsOutput, lastPage bool) bool {
-		if page == nil {
-			return !lastPage
-		}
-
-		for _, v := range page.StandardsSubscriptions {
-			if v != nil {
-				output = append(output, v)
-			}
-		}
-
-		return !lastPage
-	})
-
-	if tfawserr.ErrCodeEquals(err, securityhub.ErrCodeResourceNotFoundException) {
-		return nil, &resource.NotFoundError{
-			LastError:   err,
-			LastRequest: input,
-		}
-	}
-
-	if tfawserr.ErrMessageContains(err, securityhub.ErrCodeInvalidAccessException, "not subscribed to AWS Security Hub") {
-		return nil, &resource.NotFoundError{
-			LastError:   err,
-			LastRequest: input,
-		}
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	return output, nil
-}
-
-func FindStandardsSubscriptionByARN(ctx context.Context, conn *securityhub.SecurityHub, arn string) (*securityhub.StandardsSubscription, error) {
-	input := &securityhub.GetEnabledStandardsInput{
-		StandardsSubscriptionArns: aws.StringSlice([]string{arn}),
-	}
-
-	output, err := FindStandardsSubscription(ctx, conn, input)
-
-	if err != nil {
-		return nil, err
-	}
-
-	if status := aws.StringValue(output.StandardsStatus); status == securityhub.StandardsStatusFailed {
-		return nil, &resource.NotFoundError{
-			Message:     status,
-			LastRequest: input,
-		}
-	}
-
-	return output, nil
-}
