+++ internal/service/securityhub/status.go
@@ -6,7 +6,6 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/securityhub"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
 )
 
 const (
@@ -15,8 +14,6 @@ const (
 
 	// AdminStatus Unknown
 	adminStatusUnknown = "Unknown"
-
-	standardsStatusNotFound = "NotFound"
 )
 
 // statusAdminAccountAdmin fetches the AdminAccount and its AdminStatus
@@ -33,23 +30,5 @@ func statusAdminAccountAdmin(ctx context.Context, conn *securityhub.SecurityHub,
 		}
 
 		return adminAccount, aws.StringValue(adminAccount.Status), nil
-	}
-}
-
-func statusStandardsSubscription(ctx context.Context, conn *securityhub.SecurityHub, arn string) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		output, err := FindStandardsSubscriptionByARN(ctx, conn, arn)
-
-		if tfresource.NotFound(err) {
-			// Return a fake result and status to deal with the INCOMPLETE subscription status
-			// being a target for both Create and Delete.
-			return "", standardsStatusNotFound, nil
-		}
-
-		if err != nil {
-			return nil, "", err
-		}
-
-		return output, aws.StringValue(output.StandardsStatus), nil
 	}
 }
